requirements.txt
Django>=5.0,<5.3
psycopg[binary]>=3.1
djangorestframework>=3.15,<3.17
django-allauth>=65.0
django-environ>=0.11
django-htmx>=1.19
django-storages>=1.14
boto3>=1.35
celery>=5.4
redis>=5.0
requests>=2.32
gunicorn>=22.0
pytest>=8.3
pytest-django>=4.9
